Fix clippy lints in bridge-solver and dealer-dds
- Add Default derive and is_empty() to OrderedCards
- Collapse nested if statements in order_leads
- Use iterator instead of index loop in add_discards
- Remove unnecessary .clone() on Copy types
- Add #[allow(clippy::too_many_arguments)] where needed
- Fix empty line after doc comment in cards.rs
- Derive Default for ShapeEntry instead of manual impl
- Remove unnecessary cast in dealer-dds

+++ dealer-dds/src/lib.rs
@@ -300,7 +300,7 @@ impl GameState {
         }
 
         // Record play: high 2 bits = position, low 6 bits = card index
-        let encoded = ((player as u8) << 6) | (card.to_index() as u8);
+        let encoded = ((player as u8) << 6) | card.to_index();
         self.play_history[self.plays_count as usize] = encoded;
         self.plays_count += 1;
